requirements.txt
pynput
pygsheets
piper-tts
sounddevice